third_party/blink/web_tests/platform/generic/external/wpt/html/infrastructure/urls/resolving-urls/query-encoding/location.sub_encoding=windows-1252-expected.txt

This is a testharness.js-based test.
PASS location [PutForwards]
PASS location.assign()
PASS location.replace()
PASS location.href
FAIL location.search assert_equals: expected "?%FF" but got "?%C3%BF"
Harness: the test ran to completion.
